+++ src/schemathesis/cli/__init__.py
@@ -93,23 +93,6 @@ def register_target(function: Target) -> Target:
     return function
 
 
-def register_check(function: CheckFunction) -> CheckFunction:
-    """Register a new check for schemathesis CLI.
-
-    :param function: A function to validate API responses.
-
-    .. code-block:: python
-
-        @schemathesis.register_check
-        def new_check(response, case):
-            # some awesome assertions!
-            ...
-    """
-    checks_module.ALL_CHECKS += (function,)
-    CHECKS_TYPE.choices += (function.__name__,)  # type: ignore
-    return function
-
-
 def reset_checks() -> None:
     """Get checks list to their default state."""
     # Useful in tests